Added automated write back of wave files.
Audio data will be written with default rate 10Hz (configurable) and
appended to current file. (change file mode to read/write)
File will be named according to the start time of the node.
Additionally, now the sample vector size can be predefined, as the
sample vector is cleared (move data to new vector, which is then written
to file).
Changed default behavior to run the node until Ctrl+c is pressed to have
a continuously transporting node. In order to run the node forever, a
switch for recording should be added to avoid filling the disc (/tmp/
folder).